Removed cancel, search, update, made hpp files
