+++ HomeWork6/RafigRzayev_HW6/cafe.hpp
@@ -1,0 +1,7 @@
+#pragma once
+#include "queue.hpp"
+#include "logger.hpp"
+
+void Order(Queue &Cafe);
+void Cook(Queue &Cafe);
+void Clear(Queue &Cafe);
